fix: Move header para pasta shared de componentes
