+++ src/views/HomeView.vue
@@ -3,7 +3,7 @@ import TopicCard from '@/components/TopicCard.vue'
 import PrimaryButton from '@/components/button/PrimaryButton.vue'
 import type Topic from '@/interfaces/Topic'
 import NewTopicCard from '@/components/NewTopicCard.vue'
-import HeaderApp from '@/components/HeaderApp.vue'
+import HeaderApp from '@/components/shared/HeaderApp.vue'
 import { useTopicsStore } from '@/stores/topicStore'
 import draggable from 'vuedraggable'
 import { toast } from 'vue3-toastify'
